Completed the functionapp to serviceBuc logic and DLQ logic. Added the architecture diagram.

+++ Services/ProductService.cs
@@ -19,19 +19,28 @@ namespace ProductAPIFunctionApp.Services
 
         public async Task HandleMessageAsync(ProductMessage message)
         {
-            switch (message.Action)
+            if (message == null)
+                throw new ArgumentNullException(nameof(message));
+
+            switch (message.Action?.ToUpper())
             {
                 case "CREATE":
+                    if (message.Product == null)
+                        throw new ArgumentException("Product is required!");
+
                     message.Product.ProductCreatedDate = DateTime.UtcNow;
                     await _productRepository.CreateAsync(message.Product);
                     break;
 
                 case "DELETE":
+                    if (String.IsNullOrEmpty(message.ProductId))
+                        throw new ArgumentException("ProductId is required!");
+
                     await _productRepository.DeleteAsync(message.ProductId);
                     break;
 
                 default:
-                    throw new BadRequestException("Invalid Action");
+                    throw new ArgumentException("Invalid Action");
             }
         }
     }
